Implement gh-actions builds (#27)
