+++ .gitignore
@@ -1,15 +1,22 @@
+.idea
+.code
 **/vendor/
 .pulumi
 **/bin/
 **/obj/
 Pulumi.*.yaml
-sdk/nodejs/node_modules
+**/node_modules/
 
 **/command-output/
 
 .idea/
-.vscode/
 *.iml
 
 yarn.lock
 **/pulumiManifest.go
+
+ci-scripts
+**/schema.go
+**/version.txt
+**/nuget
+**/dist
