+++ examples/examples_dotnet_test.go
@@ -1,0 +1,32 @@
+// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+// +build dotnet all
+
+package examples
+
+import (
+	"path"
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func getCsharpBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	base := getBaseOptions(t)
+	baseCsharp := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			"Pulumi.SpotInst",
+		},
+	})
+
+	return baseCsharp
+}
+
+func TestAccAwsElastiGroupCsharp(t *testing.T) {
+	getToken(t)
+	test := getCsharpBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "aws_elastigroup", "csharp"),
+		})
+
+	integration.ProgramTest(t, &test)
+}
