+++ examples/examples_py_test.go
@@ -1,0 +1,33 @@
+// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+// +build python all
+
+package examples
+
+import (
+	"path"
+	"path/filepath"
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func getPythonBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	base := getBaseOptions(t)
+	basePython := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			filepath.Join("..", "sdk", "python", "bin"),
+		},
+	})
+
+	return basePython
+}
+
+func TestAccAwsElastiGroupPython(t *testing.T) {
+	getToken(t)
+	test := getPythonBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "aws_elastigroup", "python"),
+		})
+
+	integration.ProgramTest(t, &test)
+}
