+++ examples/examples_test.go
@@ -4,42 +4,10 @@ package examples
 
 import (
 	"os"
-	"path"
-	"path/filepath"
 	"testing"
 
 	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
 )
-
-func TestAccAwsElastiGroupTs(t *testing.T) {
-	getToken(t)
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "aws_elastigroup", "ts"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccAwsElastiGroupPython(t *testing.T) {
-	getToken(t)
-	test := getPythonBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "aws_elastigroup", "python"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccAwsElastiGroupCsharp(t *testing.T) {
-	getToken(t)
-	test := getCsharpBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "aws_elastigroup", "csharp"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
 
 func getToken(t *testing.T) {
 	token := os.Getenv("SPOTINST_TOKEN")
@@ -73,37 +41,4 @@ func getBaseOptions(t *testing.T) integration.ProgramTestOptions {
 			"aws:region": getEnvRegion(t),
 		},
 	}
-}
-
-func getJSBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	baseJS := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"@pulumi/spotinst",
-		},
-	})
-
-	return baseJS
-}
-
-func getCsharpBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	baseCsharp := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"Pulumi.SpotInst",
-		},
-	})
-
-	return baseCsharp
-}
-
-func getPythonBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	basePython := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			filepath.Join("..", "sdk", "python", "bin"),
-		},
-	})
-
-	return basePython
 }
