--- provider/cmd/pulumi-resource-spotinst/.gitignore
@@ -1,1 +1,0 @@
-schema.go
